Move Entity AI logic to its own package
