+++ src/superdopesquad/superdopejedimod/entity/StormTrooperEntity.java
@@ -41,6 +41,9 @@ import net.minecraftforge.fml.common.registry.GameRegistry;
 import net.minecraftforge.fml.relauncher.Side;
 import net.minecraftforge.fml.relauncher.SideOnly;
 import superdopesquad.superdopejedimod.SuperDopeJediMod;
+import superdopesquad.superdopejedimod.entity.ai.EntityAIEnemyFactionDetector;
+import superdopesquad.superdopejedimod.entity.ai.EntityAIHoldAndShoot;
+import superdopesquad.superdopejedimod.entity.ai.EntityAIQuicklyOffended;
 import superdopesquad.superdopejedimod.faction.ClassInfo;
 import superdopesquad.superdopejedimod.faction.ClassManager;
 import superdopesquad.superdopejedimod.faction.FactionInfo;
