+++ src/superdopesquad/superdopejedimod/entity/ai/EntityAIFollowPathMarkers.java
@@ -1,4 +1,4 @@
-package superdopesquad.superdopejedimod.entity;
+package superdopesquad.superdopejedimod.entity.ai;
 
 import java.util.List;
