+++ src/superdopesquad/superdopejedimod/entity/droid/ImperialProbeDroidEntity.java
@@ -56,11 +56,11 @@ import net.minecraftforge.fml.relauncher.Side;
 import net.minecraftforge.fml.relauncher.SideOnly;
 import superdopesquad.superdopejedimod.SuperDopeJediMod;
 import superdopesquad.superdopejedimod.entity.BaseEntityAnimal;
-import superdopesquad.superdopejedimod.entity.EntityAIEnemyFactionDetector;
-import superdopesquad.superdopejedimod.entity.EntityAIFollowPathMarkers;
-import superdopesquad.superdopejedimod.entity.EntityAIHoldAndShoot;
-import superdopesquad.superdopejedimod.entity.EntityAIQuicklyOffended;
 import superdopesquad.superdopejedimod.entity.EntityRenderFactory;
+import superdopesquad.superdopejedimod.entity.ai.EntityAIEnemyFactionDetector;
+import superdopesquad.superdopejedimod.entity.ai.EntityAIFollowPathMarkers;
+import superdopesquad.superdopejedimod.entity.ai.EntityAIHoldAndShoot;
+import superdopesquad.superdopejedimod.entity.ai.EntityAIQuicklyOffended;
 import superdopesquad.superdopejedimod.faction.FactionInfo;
 import superdopesquad.superdopejedimod.weapon.PlasmaShotEntityBase.PowerLevel;
 import superdopesquad.superdopejedimod.faction.ClassManager;
